+++ core/models/leavemsg.go
@@ -12,7 +12,7 @@ type LeaveMsg struct {
 	UpdatedAt time.Time      `gorm:"column:updated_at;type:timestamp;not null;default:CURRENT_TIMESTAMP;autoUpdateTime"`
 	DeletedAt gorm.DeletedAt `gorm:"column:deleted_at;type:timestamp;index"`
 
-	MsgID    int64  `gorm:"column:msg_id;type:bigint;not null;uniqueIndex"`
+	MsgID    uint32 `gorm:"column:msg_id;type:bigint;not null;uniqueIndex"`
 	Content  string `gorm:"column:content;type:text;not null"`
 	AuthorID int64  `gorm:"column:author_id;type:bigint;not null;index"`
